tech: extract useRunEntity logic
